search and add work
searches.txt now works for storing search items and titles
the add search item option now works. after an item is added, the loop is reset, data is reloaded and the searching starts over.
need to plan delete function as well as ad a delete button to the "x item found" dialog so you can delete on the go.